Move log/trace context functions to individual packages
This allows moving log-context related code into the log package as
this is a better location for it.

Moving trace out is required to avoid an import cycle:

log -> context (for trace) -> log (for Gorm logger)

Technically, I think the coupling between logger and trace should
probably move to a separate piece of middleware, that combines the two
elements, something like `LogWithTrace`.