credentials: Fix Vault autorenewal to not deadlock when renewal fails
The Vault credential renewal goroutine will deadlock when a token
becomes un-renewable, because the code is `break`ing out of the
`select` used to distinguish between renewal success and failure.

So instead of attempting to re-authenticate when the renewal failed,
the code was waiting forever for another renewal success/failure
signal that would never be sent by the (now exited) Vault client
renewal goroutine, and would thus deadlock.

The fix is to break out of the `for` loop waiting for renewals instead
of the `select`.

---

Background: Vault has two mechanisms for lifetime management:

* TTLs: A token has a TTL and may have a max TTL. When renewing a
  token, the TTL can be set to some interval, and the token's expiry
  will be pushed out from renewal time to renewal time + TTL. *But* a
  token *cannot* be extended beyond the initial token creation date +
  max TTL. If renewing the token will push the token lifetime beyond
  the max TTL, then the Vault client will terminate the background
  renewal goroutine.

* Periods: This allows tokens to avoid the lifetime limit, by allowing
  tokens to be perpetually renewed as long as they are renewed within
  a given period. So if a token has a period of 1h, it can be renewed
  as long as it was either created or renewed within 1h of now.

https://www.vaultproject.io/docs/concepts/tokens.html